+++ activerecord/lib/active_record/connection_adapters/postgresql/schema_statements.rb
@@ -448,7 +448,7 @@ module ActiveRecord
           end
         end
 
-        def remove_index(table_name, column_name = nil, options = {}) #:nodoc:
+        def remove_index(table_name, column_name = nil, options = {}) # :nodoc:
           table = Utils.extract_schema_qualified_name(table_name.to_s)
 
           if column_name.is_a?(Hash)
@@ -467,13 +467,17 @@ module ActiveRecord
             end
           end
 
+          return if options[:if_exists] && !index_exists?(table_name, column_name, options)
+
           index_to_remove = PostgreSQL::Name.new(table.schema, index_name_for_remove(table.to_s, column_name, options))
+
           algorithm =
             if options.key?(:algorithm)
               index_algorithms.fetch(options[:algorithm]) do
                 raise ArgumentError.new("Algorithm must be one of the following: #{index_algorithms.keys.map(&:inspect).join(', ')}")
               end
             end
+
           execute "DROP INDEX #{algorithm} #{quote_table_name(index_to_remove)}"
         end
